Refactor renderTodolist into multiple modules
