+++ src/_test_/test_fles/constructors.test.js
@@ -1,16 +1,16 @@
-import { Project, Task } from '../functions/constructors';
+import * as constructors from "../../assets/js/constructors";
 
-test('Expect Project constructor to create a new project object', () => {
-  const newProject = new Project('Some Project');
-  expect(newProject).toEqual({ name: 'Some Project', tasks: [] });
+test("Expect Project constructor to create a new project object", () => {
+  const newProject = new constructors.Project("Some Project");
+  expect(newProject).toEqual({ name: "Some Project", tasks: [] });
 });
 
-test('Expect Task constructor to create a new task object', () => {
-  const newTask = new Task('Task 1', 'High', '30-12-2020', [1, 1]);
+test("Expect Task constructor to create a new task object", () => {
+  const newTask = new constructors.Task("Task 1", "High", "30-12-2020", [1, 1]);
   expect(newTask).toEqual({
-    title: 'Task 1',
-    priority: 'High',
-    date: '30-12-2020',
+    title: "Task 1",
+    priority: "High",
+    date: "30-12-2020",
     id: [1, 1],
     status: true,
   });
