+++ src/_test_/test_fles/helpers.test.js
@@ -1,23 +1,19 @@
-import {
-  addProjectCount,
-  addTaskCount,
-  getProjectIndex,
-} from '../functions/helpers';
+import * as helpers from "../../assets/js/helpers";
 
-test('Expect addProjectCount to increase project count by 1', () => {
-  localStorage.setItem('projectCounter', '0');
+test("Expect addProjectCount to increase project count by 1", () => {
+  localStorage.setItem("projectCounter", "0");
 
-  expect(addProjectCount()).toEqual(1);
+  expect(helpers.addProjectCount()).toEqual(1);
 });
 
-test('Expect addTaskCount to increase task count by 1', () => {
-  localStorage.setItem('taskCounter', '0');
+test("Expect addTaskCount to increase task count by 1", () => {
+  localStorage.setItem("taskCounter", "0");
 
-  expect(addTaskCount()).toEqual(1);
+  expect(helpers.addTaskCount()).toEqual(1);
 });
 
-test('Expect getProjectIndex to get the project ID from project element', () => {
+test("Expect getProjectIndex to get the project ID from project element", () => {
   const fakeElement = { dataset: 5 };
 
-  expect(getProjectIndex(fakeElement)).toEqual(5);
+  expect(helpers.getProjectIndex(fakeElement)).toEqual(5);
 });
